perf(stacks): split StackResourceItem into per-tab subcomponents
The 1660-line StackResourceItem render function caused every keystroke in
Resource Name to re-render the whole Configuration / Deployment /
Environment tree, including all Radix Tooltips, Selects, env-row
groupings. React.memo on the parent didn't help because the resource
prop changed per keystroke.

Split into a thin parent (~330 lines) plus three React.memo'd children:
- StackResourceConfigurationTab — General + Source/Build sections
- StackResourceDeploymentTab — Init/Main command + args + ports
- StackResourceEnvironmentTab — env vars table + addon groups

Each child receives a per-tab projected draft (pickConfigurationDraft,
pickDeploymentDraft) and stable patch callbacks held in refs. Typing in
Resource Name now only re-renders Configuration; Deployment and
Environment stay idle.

Trace numbers (30-keystroke burst on resource-name-2):
- per-key:        52ms  -> 38.8ms (-25%)
- longest task:   178ms -> 94.5ms (-47%)  -- exit criterion < 100ms hit
- long tasks >50: 2     -> 2 (unchanged, but each is much shorter)

Behaviour preserved: per-field DirtyField tints + reset arrows, tab
dirty dots, Modified pill with discard, all callbacks (onChange,
onRemove, onDiscardField, onDiscardEnvRow, onDiscardResource), accordion
header, Tabs structure, and the outer TooltipProvider scope.
